Convert useRegistry to TS
